# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Les Canetons API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "https://lescanetons.org/api/v1",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/me/password": {
            "post": {
                "operationId": "account.password",
                "description": "Any logged-in account holder, for their own account. No permission is\nneeded: this is the screen every member has and nobody administers, and\nit is where a member sent by `mustChangePassword` lands.\n\nSend `currentPassword` and `newPassword`. The new one must be at least\neight characters.\n\nOn success the password is replaced, the forced-change flag is cleared,\nand every other session the member has open is ended. Answers\n`{\"ok\": true, \"sessionsEnded\": n}`, where `sessionsEnded` counts those\nother sessions; the calling session stays valid.\n\nA wrong `currentPassword` answers `403 reauth_failed`. After five wrong\nones the account answers `429 too_many_attempts` for fifteen minutes,\nand a correct password during that window is still refused. A\n`newPassword` shorter than eight characters answers\n`400 validation_failed` with `too_short` against `newPassword`.",
                "summary": "Change your own password",
                "tags": [
                    "Session"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "sessionsEnded": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "sessionsEnded"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
        "/events/{event}": {
            "get": {
                "operationId": "event.show",
                "description": "Any logged-in member; no permission is needed. Returns the event with\n`myAttendance`, the caller's own answer or `null`.\n\nWorks for a past event as well as an upcoming one, unlike the default\nlist. An unknown id answers `404`.",
                "summary": "Read one event",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "event.update",
                "description": "Requires `events.manage`. Send only the fields that change: an omitted\nfield is left alone, and an explicit `null` clears an optional one such\nas `attire` or `notes`. Returns the updated event, including the\ncaller's own `myAttendance`.\n\n`endsAt` must still come after the start. A request that sends a new end\nand no new start is compared against the stored `startsAt`, and fails\nagainst `endsAt` with `must_be_after` if it falls before it.\n\nClearing `registrationClosesAt` switches public registration off again.",
                "summary": "Correct an event already on the planning",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`EventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "delete": {
                "operationId": "event.destroy",
                "description": "Requires `events.manage`. Deletes the event and everything attached to\nit: every attendance answer given for it, and every public registration\nbooked for it.\n\nAnswers `{\"ok\": true, \"attendanceDeleted\": n, \"registrationsDeleted\": n}`\nwith the counts of what went with it, so a confirmation can say what is\nabout to be discarded rather than merely asking again.",
                "summary": "Take an event off the planning",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "attendanceDeleted": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "registrationsDeleted": {
                                            "type": "integer",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "attendanceDeleted",
                                        "registrationsDeleted"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
        "/events/{event}/attendance": {
            "put": {
                "operationId": "attendance.update",
                "description": "Any logged-in member who is in a register; no permission is needed.\nAnswering is idempotent: sending an answer again replaces the previous\none rather than adding a second, and the response is always `200` with\nthe answer as it now stands, including its `recordedAt`.\n\n`status` is `yes` or `no`; anything else fails validation against that\nfield with `invalid_value`, which lists what is accepted.\n\nTaking back a `yes` costs a reason: if the stored answer is `yes` and\nthe incoming one is `no`, `note` is required, and a blank one fails\nvalidation against `note` with `required`. A first answer of `no`, and\na change from `no` to `yes`, cost nothing. Recording on somebody else's\nbehalf is exempt from this rule.\n\nA member who is in no register is refused with `403 not_answerable`:\nnothing is being asked of them, and no permission would change that.\n\nAnswering for yourself clears any mark saying the direction entered the\nanswer.",
                "summary": "Record your own answer for an event",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordOwnAttendanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AttendanceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttendanceResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "delete": {
                "operationId": "attendance.destroy",
                "description": "Any logged-in member; no permission is needed. Removes the answer\nentirely and returns the event to unanswered, which is a state a second\n`PUT` cannot express. Answers `{\"ok\": true}`, and taking back an answer\nthat is not there is not an error.\n\nThe window closes five minutes after the answer was last recorded;\nafter that it answers `409 answer_already_settled`. `recordedAt` on the\nanswer is what tells a client whether the window is still open, so it\ncan offer the undo or not rather than finding out from a refusal.",
                "summary": "Take back your own answer",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "get": {
                "operationId": "attendance.index",
                "description": "Requires `attendance.view_all`. Returns every member who is answerable\nfor the event, ordered by name, whether or not they have replied: an\nentry's `attendance` is `null` for somebody who has not, which is what\nthis list is read for.\n\nAnswerable means being in a register. A member who is in none, such as\nsomebody who only organises, never appears here.\n\nEach entry carries the member's name and register alongside their\nanswer, and an answer says whether the direction entered it rather than\nthe member giving it themselves.",
                "summary": "List who is coming to an event",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `ChaseListEntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChaseListEntryResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "operationId": "role.index",
                "description": "Requires `members.manage`. Read-only reference data, for a client\nbuilding the roster form.\n\nReturns every role with the permissions it grants. A role carries no\ndisplay name: `key` is the fixed identifier and the client translates\nit.",
                "summary": "List the roles",
                "tags": [
                    "Members"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `RoleResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoleResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    }
                }
            }
        },
...
        "/members/{member}/password": {
            "post": {
                "operationId": "memberPassword.reset",
                "description": "Requires `members.manage`. Mints a readable password an administrator\ncan dictate over the phone, replaces the member's own with it, and ends\nevery session that member has open. Takes no body. Answers\n`{\"generatedPassword\": \"...\", \"sessionsEnded\": n}`.\n\n`generatedPassword` is shown in this response and nowhere else: it is\nhashed on the way into the database, is never written to the audit log,\nand no later call returns it. A lost one is replaced by calling this\nagain. The member is required to change it at their next login, through\n`POST /api/v1/me/password`.\n\nA member resetting their own password this way keeps the session they\nare calling from, and `sessionsEnded` then counts their other ones. The\nordinary route for that is `POST /api/v1/me/password`, which lets them\nchoose the password instead.",
                "summary": "Reset a member's password",
                "tags": [
                    "Members"
                ],
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "generatedPassword": {
                                            "type": "string"
                                        },
                                        "sessionsEnded": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "generatedPassword",
                                        "sessionsEnded"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountPasswordRequest": {
                "type": "object",
                "description": "The credentials for changing your own password.\n\nSend the password the account is using now and the one to replace it with.\nThe current password is checked again server-side before anything is\nwritten, so knowing it is what separates a borrowed session from a\npermanently stolen account.\n\nThe only requirement on the new password is a minimum of eight characters.\nThere is no complexity rule, no reuse check and no expiry.",
                "properties": {
                    "currentPassword": {
                        "type": "string",
                        "description": "The password the account is using right now. Re-verified before the change is applied; a wrong one answers `403 reauth_failed`."
                    },
                    "newPassword": {
                        "type": "string",
                        "description": "The replacement. At least 8 characters; a shorter one fails with `too_short`.",
                        "minLength": 8,
                        "maxLength": 255
                    }
                },
                "required": [
                    "currentPassword",
                    "newPassword"
                ],
                "title": "AccountPasswordRequest"
            },
...
        },
        "responses": {
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}